# Fix internal sums of QArrays