--- a/dynamiqs/utils/vectorization.py
+++ b/dynamiqs/utils/vectorization.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-import jax
 import numpy as np
 
 from .._checks import check_shape
 from ..qarrays.qarray import QArray, QArrayLike
-from ..qarrays.utils import asqarray
+from ..qarrays.utils import asqarray, sum_qarray_pytree
 from .operators import eye
 from .quantum_utils import dag
 
@@ -251,12 +250,4 @@ def slindbladian(H: QArrayLike, jump_ops: list[QArrayLike]) -> QArray:
         check_shape(L, f'jump_ops[{i}]', '(..., n, n)')
 
     terms = [-1j * spre(H), 1j * spost(H)] + [sdissipator(L) for L in jump_ops]
-    return sum_qarrays(terms)
-
-
-def sum_qarrays(qarrays: list[QArray]) -> QArray:
-    # jax.tree.reduce doesn't call its initializer if the list is not empty
-    # this avoids unwanted conversion from sparse to dense qarrays when summing with 0
-    return jax.tree.reduce(
-        lambda x, y: x + y, qarrays, is_leaf=lambda x: isinstance(x, QArray)
-    )
+    return sum_qarray_pytree(terms)