Proper fix for 74b18cee880b30dd61773b9155a3c7924e832f45.

Also fixes the implementation of `vector_field` for `dq.mesolve`. This yields 5-10% improvement on small systems (and equivalent runtimes on larger systems), e.g. on:
```python
import jax
import jax.numpy as jnp
import dynamiqs as dq

# === Prepare objects ===
# parameters
n = 50
delta = 0.1
epsilon = 2
gamma = 1
nth = 0.8

# operators
a = dq.destroy(n)
H = delta * dq.dag(a) @ a + epsilon * (a + dq.dag(a))
L = [jnp.sqrt(gamma * (1 + nth)) * a, jnp.sqrt(gamma * nth) * dq.dag(a)]
y = dq.todm(dq.fock(n, 0))

# save times
tsave = jnp.linspace(0.0, 1.0, 101)
options = dq.Options(progress_meter=None)
dq.mesolve(H, L, y, tsave, options=options).block_until_ready()
%timeit dq.mesolve(H, L, y, tsave, options=options).block_until_ready()
```
Before: `43 ms ± 344 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)`
After: `37.6 ms ± 266 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)`
(reproducible over many tries)

Benchmark difference can also be checked directly on the vector field definition:
```python
import jax
import jax.numpy as jnp
import dynamiqs as dq
from dynamiqs.qarrays.utils import tree_sum

# === Prepare objects ===
# parameters
n = 50
delta = 0.1
epsilon = 2
gamma = 1
nth = 0.8

# operators
a = dq.destroy(n)
H = delta * dq.dag(a) @ a + epsilon * (a + dq.dag(a))
L = [jnp.sqrt(gamma * (1 + nth)) * a, jnp.sqrt(gamma * nth) * dq.dag(a)]
y = dq.todm(dq.fock(n, 0))

# === Vector field definition ===
@jax.jit
def vector_field_old(H, L, y):
    L = dq.stack(L)
    Ld = L.dag()
    LdL = (Ld @ L).sum(0)
    tmp = (-1j * H - 0.5 * LdL) @ y + 0.5 * (L @ y @ Ld).sum(0)
    return tmp + tmp.dag()

@jax.jit
def vector_field_new(H, L, y):
    Hnh = tree_sum([-1j * H] + [-0.5 * _L.dag() @ _L for _L in L])
    tmp = tree_sum([Hnh @ y] + [0.5 * _L @ y @ _L.dag() for _L in L])
    return tmp + tmp.dag()


# === Run ===
vector_field_old(H, L, y).block_until_ready()
%timeit vector_field_old(H, L, y).block_until_ready()

vector_field_new(H, L, y).block_until_ready()
%timeit vector_field_new(H, L, y).block_until_ready()
```
Before: `109 µs ± 1.96 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)`
After: `95.5 µs ± 2.29 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)`